Extract string literal reference base class
